wip: adds an in-memory cache to curl requests
Caches repeated requests via curl with an in-memory cache.
This allows the tests to make each request only once, and more tests can
be written without encroaching further on third-party service
rate-limits.

Note that this in-memory cache lives for the life of the program, so
long-running services that use neil as a library would need to
periodically clear the cache, unless some other clearing mechanism is
implemented.

In this model, typical bb one-off usages should always start with an
empty cache, so there should be little change for end-users.